+++ glusterd2/commands/command.go
@@ -2,7 +2,6 @@
 package commands
 
 import (
-	"github.com/gluster/glusterd2/glusterd2/commands/devices"
 	"github.com/gluster/glusterd2/glusterd2/commands/global"
 	"github.com/gluster/glusterd2/glusterd2/commands/peers"
 	"github.com/gluster/glusterd2/glusterd2/commands/version"
@@ -23,6 +22,5 @@ var Commands = []Command{
 	&versioncommands.Command{},
 	&volumecommands.Command{},
 	&peercommands.Command{},
-	&devicecommands.Command{},
 	&globalcommands.Command{},
 }
