+++ plugins/device/api/req.go
@@ -1,4 +1,4 @@
-package api
+package device
 
 const (
 	// DeviceEnabled represents enabled
@@ -16,5 +16,3 @@ type AddDeviceReq struct {
 	Devices []string `json:"devices"`
 }
 
-// DeviceAddResp is the success response sent to a AddDeviceReq request
-type DeviceAddResp Peer
