+++ plugins/device/api/res.go
@@ -1,0 +1,9 @@
+package device
+
+import (
+	"github.com/gluster/glusterd2/pkg/api"
+)
+
+// DeviceAddResp is the success response sent to a AddDeviceReq request
+type DeviceAddResp api.Peer
+
