+++ plugins/device/init.go
@@ -1,0 +1,43 @@
+package device
+
+import (
+        "github.com/gluster/glusterd2/glusterd2/servers/rest/route"
+        "github.com/gluster/glusterd2/glusterd2/transaction"
+	deviceapi "github.com/gluster/glusterd2/plugins/device/api"
+        "github.com/gluster/glusterd2/pkg/utils"
+	"github.com/gluster/glusterd2/pkg/sunrpc"
+)
+
+// Plugin is a structure which implements GlusterdPlugin interface
+type Plugin struct {
+}
+
+// Name returns name of plugin
+func (p *Plugin) Name() string {
+        return "georeplication"
+}
+
+// SunRPCProgram returns sunrpc program to register with Glusterd
+func (p *Plugin) SunRPCProgram() sunrpc.Program {
+        return nil
+}
+
+// Routes returns list of REST API routes to register with Glusterd.
+func (p *Plugin) RestRoutes() route.Routes {
+        return route.Routes{
+                route.Route{
+                        Name:         "DeviceAdd",
+                        Method:       "POST",
+                        Pattern:      "/peers/{peerid}/devices",
+                        Version:      1,
+                        RequestType:  utils.GetTypeString((*deviceapi.AddDeviceReq)(nil)),
+                        ResponseType: utils.GetTypeString((*deviceapi.DeviceAddResp)(nil)),
+                        HandlerFunc:  deviceAddHandler},
+        }
+}
+
+// RegisterStepFuncs registers transaction step functions with
+// Glusterd Transaction framework
+func (p *Plugin) RegisterStepFuncs() {
+        transaction.RegisterStepFunc(txnPrepareDevice, "prepare-device")
+}
