+++ plugins/device/transaction.go
@@ -1,12 +1,12 @@
-package devicecommands
+package device
 
 import (
 	"os/exec"
 	"strings"
 
-	"github.com/gluster/glusterd2/glusterd2/device"
 	"github.com/gluster/glusterd2/glusterd2/transaction"
 	"github.com/gluster/glusterd2/pkg/api"
+	deviceapi "github.com/gluster/glusterd2/plugins/device/api"
 
 	"github.com/pborman/uuid"
 	log "github.com/sirupsen/logrus"
@@ -14,7 +14,7 @@ import (
 
 func txnPrepareDevice(c transaction.TxnCtx) error {
 	var peerID uuid.UUID
-	var req api.AddDeviceReq
+	var req deviceapi.AddDeviceReq
 	var deviceList []api.DeviceInfo
 	if err := c.Get("peerid", peerID); err != nil {
 		c.Logger().WithError(err).Error("Failed transaction, cannot find peer-id")
@@ -34,18 +34,18 @@ func txnPrepareDevice(c transaction.TxnCtx) error {
 		pvcreateCmd := exec.Command("pvcreate", "--metadatasize=128M", "--dataalignment=256K", element.Name)
 		if err := pvcreateCmd.Run(); err != nil {
 			c.Logger().WithError(err).WithField("device", element.Name).Error("pvcreate failed for device")
-			deviceList[index].State = api.DeviceFailed
+			deviceList[index].State = deviceapi.DeviceFailed
 			continue
 		}
 		vgcreateCmd := exec.Command("vgcreate", strings.Replace("vg"+element.Name, "/", "-", -1), element.Name)
 		if err := vgcreateCmd.Run(); err != nil {
 			c.Logger().WithError(err).WithField("device", element.Name).Error("vgcreate failed for device")
-			deviceList[index].State = api.DeviceFailed
+			deviceList[index].State = deviceapi.DeviceFailed
 			continue
 		}
-		deviceList[index].State = api.DeviceEnabled
+		deviceList[index].State = deviceapi.DeviceEnabled
 	}
-	err := device.AddDevices(deviceList, peerID.String())
+	err := AddDevices(deviceList, peerID.String())
 	if err != nil {
 		log.WithError(err).Error("Couldn't add deviceinfo to store")
 	}
